Move keyboard to separate file.
