+++ vt05.c
@@ -19,11 +19,8 @@
 #include <SDL.h>
 #include <assert.h>
 
+#include "terminal.h"
 #include "args.h"
-
-typedef uint32_t u32;
-typedef uint8_t u8;
-#define nil NULL
 
 SDL_Surface *screen;
 
@@ -363,216 +360,6 @@ recvchar(int c)
 	updatebuf = 1;
 }
 
-/* Map SDL scancodes to ASCII */
-// Can't map VT05 keyboard sensible....this is from Datapoint 3300
-int scancodemap_orig[SDL_NUM_SCANCODES] = {
-	[SDL_SCANCODE_1] = '1',
-	[SDL_SCANCODE_2] = '2',
-	[SDL_SCANCODE_3] = '3',
-	[SDL_SCANCODE_4] = '4',
-	[SDL_SCANCODE_5] = '5',
-	[SDL_SCANCODE_6] = '6',
-	[SDL_SCANCODE_7] = '7',
-	[SDL_SCANCODE_8] = '8',
-	[SDL_SCANCODE_9] = '9',
-	[SDL_SCANCODE_0] = '0',
-	[SDL_SCANCODE_MINUS] = ':',
-	[SDL_SCANCODE_EQUALS] = '-',
-	[SDL_SCANCODE_BACKSPACE] = 010,
-	[SDL_SCANCODE_DELETE] = 0177,
-
-	[SDL_SCANCODE_ESCAPE] = 033,
-	[SDL_SCANCODE_TAB] = 033,	/* Or map this to tab */
-	[SDL_SCANCODE_Q] = 'Q',
-	[SDL_SCANCODE_W] = 'W',
-	[SDL_SCANCODE_E] = 'E',
-	[SDL_SCANCODE_R] = 'R',
-	[SDL_SCANCODE_T] = 'T',
-	[SDL_SCANCODE_Y] = 'Y',
-	[SDL_SCANCODE_U] = 'U',
-	[SDL_SCANCODE_I] = 'I',
-	[SDL_SCANCODE_O] = 'O',
-	[SDL_SCANCODE_P] = 'P',
-	[SDL_SCANCODE_LEFTBRACKET] = 012,
-	[SDL_SCANCODE_RIGHTBRACKET] = 015,
-
-	[SDL_SCANCODE_A] = 'A',
-	[SDL_SCANCODE_S] = 'S',
-	[SDL_SCANCODE_D] = 'D',
-	[SDL_SCANCODE_F] = 'F',
-	[SDL_SCANCODE_G] = 'G',
-	[SDL_SCANCODE_H] = 'H',
-	[SDL_SCANCODE_J] = 'J',
-	[SDL_SCANCODE_K] = 'K',
-	[SDL_SCANCODE_L] = 'L',
-	[SDL_SCANCODE_SEMICOLON] = ';',
-	[SDL_SCANCODE_APOSTROPHE] = 0177,
-	[SDL_SCANCODE_RETURN] = 015,
-
-	[SDL_SCANCODE_Z] = 'Z',
-	[SDL_SCANCODE_X] = 'X',
-	[SDL_SCANCODE_C] = 'C',
-	[SDL_SCANCODE_V] = 'V',
-	[SDL_SCANCODE_B] = 'B',
-	[SDL_SCANCODE_N] = 'N',
-	[SDL_SCANCODE_M] = 'M',
-	[SDL_SCANCODE_COMMA] = ',',
-	[SDL_SCANCODE_PERIOD] = '.',
-	[SDL_SCANCODE_SLASH] = '/',
-	[SDL_SCANCODE_SPACE] = ' ',
-};
-
-char *scancodemap[SDL_NUM_SCANCODES] = {
-	[SDL_SCANCODE_ESCAPE] = "\033\033",
-
-	[SDL_SCANCODE_GRAVE] = "\033\033",
-	[SDL_SCANCODE_1] = "1!",
-	[SDL_SCANCODE_2] = "2@",
-	[SDL_SCANCODE_3] = "3#",
-	[SDL_SCANCODE_4] = "4$",
-	[SDL_SCANCODE_5] = "5%",
-	[SDL_SCANCODE_6] = "6^",
-	[SDL_SCANCODE_7] = "7&",
-	[SDL_SCANCODE_8] = "8*",
-	[SDL_SCANCODE_9] = "9(",
-	[SDL_SCANCODE_0] = "0)",
-	[SDL_SCANCODE_MINUS] = "-_",
-	[SDL_SCANCODE_EQUALS] = "=+",
-	[SDL_SCANCODE_BACKSPACE] = "\b\b",
-	[SDL_SCANCODE_DELETE] = "\177\177",
-
-	[SDL_SCANCODE_TAB] = "\011\011",
-	[SDL_SCANCODE_Q] = "QQ",
-	[SDL_SCANCODE_W] = "WW",
-	[SDL_SCANCODE_E] = "EE",
-	[SDL_SCANCODE_R] = "RR",
-	[SDL_SCANCODE_T] = "TT",
-	[SDL_SCANCODE_Y] = "YY",
-	[SDL_SCANCODE_U] = "UU",
-	[SDL_SCANCODE_I] = "II",
-	[SDL_SCANCODE_O] = "OO",
-	[SDL_SCANCODE_P] = "PP",
-	[SDL_SCANCODE_LEFTBRACKET] = "[[",
-	[SDL_SCANCODE_RIGHTBRACKET] = "]]",
-	[SDL_SCANCODE_BACKSLASH] = "\\\\",
-
-	[SDL_SCANCODE_A] = "AA",
-	[SDL_SCANCODE_S] = "SS",
-	[SDL_SCANCODE_D] = "DD",
-	[SDL_SCANCODE_F] = "FF",
-	[SDL_SCANCODE_G] = "GG",
-	[SDL_SCANCODE_H] = "HH",
-	[SDL_SCANCODE_J] = "JJ",
-	[SDL_SCANCODE_K] = "KK",
-	[SDL_SCANCODE_L] = "LL",
-	[SDL_SCANCODE_SEMICOLON] = ";:",
-	[SDL_SCANCODE_APOSTROPHE] = "'\"",
-	[SDL_SCANCODE_RETURN] = "\015\015",
-
-	[SDL_SCANCODE_Z] = "ZZ",
-	[SDL_SCANCODE_X] = "XX",
-	[SDL_SCANCODE_C] = "CC",
-	[SDL_SCANCODE_V] = "VV",
-	[SDL_SCANCODE_B] = "BB",
-	[SDL_SCANCODE_N] = "NN",
-	[SDL_SCANCODE_M] = "MM",
-	[SDL_SCANCODE_COMMA] = ",<",
-	[SDL_SCANCODE_PERIOD] = ".>",
-	[SDL_SCANCODE_SLASH] = "/?",
-	[SDL_SCANCODE_SPACE] = "  ",
-};
-
-int ctrl;
-int shift;
-
-void
-keydown(SDL_Keysym keysym, int repeat)
-{
-	char *keys;
-	int key;
-
-	switch(keysym.scancode){
-	case SDL_SCANCODE_LSHIFT:
-	case SDL_SCANCODE_RSHIFT: shift = 1; return;
-	case SDL_SCANCODE_CAPSLOCK:
-	case SDL_SCANCODE_LCTRL:
-	case SDL_SCANCODE_RCTRL: ctrl = 1; return;
-	}
-	if(keystate[SDL_SCANCODE_LGUI] || keystate[SDL_SCANCODE_RGUI])
-		return;
-
-	if(keysym.scancode == SDL_SCANCODE_F11 && !repeat){
-		u32 f = SDL_GetWindowFlags(window) &
-			SDL_WINDOW_FULLSCREEN_DESKTOP;
-		SDL_SetWindowFullscreen(window,
-			f ? 0 : SDL_WINDOW_FULLSCREEN_DESKTOP);
-	}
-
-/*
-	key = scancodemap[keysym.scancode];
-	if(key == 0)
-		return;
-	if(shift)
-		key ^= 020;
-	if(ctrl)
-		key &= 037;
-*/
-	keys = scancodemap[keysym.scancode];
-	if(keys == nil)
-		return;
-	key = keys[shift];
-	if(ctrl)
-		key &= 037;
-//	printf("%o(%d %d) %c\n", key, shift, ctrl, key);
-
-	char c = key;
-	write(pty, &c, 1);
-
-/*
-SDL_Event ev;
-SDL_memset(&ev, 0, sizeof(SDL_Event));
-ev.type = userevent;
-recvchar(c);
-SDL_PushEvent(&ev);
-*/
-}
-
-void
-keyup(SDL_Keysym keysym)
-{
-	switch(keysym.scancode){
-	case SDL_SCANCODE_LSHIFT:
-	case SDL_SCANCODE_RSHIFT: shift = 0; return;
-	case SDL_SCANCODE_CAPSLOCK:
-	case SDL_SCANCODE_LCTRL:
-	case SDL_SCANCODE_RCTRL: ctrl = 0; return;
-	}
-}
-
-void*
-readthread(void *p)
-{
-	char c;
-	SDL_Event ev;
-	static struct timespec slp;
-
-	SDL_memset(&ev, 0, sizeof(SDL_Event));
-	ev.type = userevent;
-
-	while(1){
-		if(read(pty, &c, 1) < 0)
-			exit(0);
-		recvchar(c);
-
-		/* simulate baudrate, TODO: get from tty */
-//		slp.tv_nsec = 1000*1000*1000 / (baud/11);
-//		nanosleep(&slp, NULL);
-
-//printf("push userevent\n");
-		SDL_PushEvent(&ev);
-	}
-}
-
 #if 0
 void*
 timethread(void *arg)
@@ -620,6 +407,8 @@ main(int argc, char *argv[])
 	int x, y;
 	pthread_t thr1, thr2;
 	struct winsize ws;
+
+	scancodemap = scancodemap_upper;
 
 	ARGBEGIN{
 	case 'b':
